requirements.txt
alembic==1.3.2
aniso8601==8.0.0
attrs==20.1.0
Babel==2.8.0
bleach==3.1.0
bleach-whitelist==0.0.10
blinker==1.4
bouncer==0.1.12
Click==7.0
dotenv-config==0.1.4
feedgen==0.9.0
Flask==1.1.1
Flask-Admin==1.5.6
Flask-Babel==1.0.0
flask-bouncer==0.3.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
flask-marshmallow==0.13.0
Flask-Migrate==2.5.2
Flask-Paranoid==0.2.0
flask-restx==0.2.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.2.0
lxml==4.5.1
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==3.7.1
marshmallow-sqlalchemy==0.23.1
psycopg2==2.8.4
PyJWT==1.7.1
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
python-slugify==4.0.0
pytz==2020.1
six==1.13.0
SQLAlchemy==1.3.12
text-unidecode==1.3
webencodings==0.5.1
Werkzeug==0.16.0
WTForms==2.2.1
